Stripping out worker from web app. Using worker to format in blocklycompiler
