+++ pxtblocks/blocklycompiler.ts
@@ -1,5 +1,6 @@
 ///<reference path='../localtypings/blockly.d.ts'/>
 /// <reference path="../built/pxtlib.d.ts" />
+/// <reference path="../built/worker.d.ts" />
 
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -8,7 +9,20 @@
 
 import B = Blockly;
 
+let iface: pxt.worker.Iface
+
 namespace pxt.blocks {
+
+    export function initWorker() {
+        if (!iface) {
+            iface = pxt.worker.makeWebWorker(pxt.webConfig.workerjs)
+        }
+    }
+
+    export function workerOpAsync(op: string, arg: pxtc.service.OpArg) {
+        initWorker()
+        return iface.opAsync(op, arg)
+    }
 
     export enum NT {
         Prefix, // op + map(children)
@@ -1340,7 +1354,7 @@ namespace pxt.blocks {
         return e;
     }
 
-    export function compileBlock(b: B.Block, blockInfo: pxtc.BlocksInfo): BlockCompilationResult {
+    export function compileBlock(b: B.Block, blockInfo: pxtc.BlocksInfo): Promise<BlockCompilationResult> {
         const w = b.workspace;
         const e = mkEnv(w, blockInfo);
         infer(e, w);
@@ -1478,7 +1492,7 @@ namespace pxt.blocks {
         return undefined;
     }
 
-    export function compile(b: B.Workspace, blockInfo: pxtc.BlocksInfo): BlockCompilationResult {
+    export function compile(b: B.Workspace, blockInfo: pxtc.BlocksInfo): Promise<BlockCompilationResult> {
         return tdASTtoTS(compileWorkspace(b, blockInfo));
     }
 
@@ -1520,7 +1534,7 @@ namespace pxt.blocks {
         ".": 18,
     }
 
-    function tdASTtoTS(app: JsNode[]): BlockCompilationResult {
+    function tdASTtoTS(app: JsNode[]): Promise<BlockCompilationResult> {
         let sourceMap: SourceInterval[] = [];
         let output = ""
         let indent = ""
@@ -1592,13 +1606,13 @@ namespace pxt.blocks {
         if (!output)
             output += "\n"
 
-        // outformat
-        //TODO: samelh output = pxtc.format(output, 1).formatted;
-
-        return {
-            source: output,
-            sourceMap: sourceMap
-        };
+        // outformat async
+        return workerOpAsync("format", { format: {input: output, pos: 1} }).then(() => {
+            return {
+                source: output,
+                sourceMap: sourceMap
+            };
+        })
 
         function emit(n: JsNode) {
             if (n.glueToBlock) {
