+++ webapp/src/app.tsx
@@ -844,7 +844,7 @@ export class ProjectView
         this.setFile(f);
     }
 
-    saveBlocksToTypeScript(): string {
+    saveBlocksToTypeScript(): Promise<string> {
         return this.blocksEditor.saveToTypeScript();
     }
 
@@ -855,20 +855,20 @@ export class ProjectView
         let promise = Promise.resolve().then(() => {
             return open ? this.textEditor.loadMonacoAsync() : Promise.resolve();
         }).then(() => {
-            let src = this.editor.saveToTypeScript();
+            return this.editor.saveToTypeScript().then((src) => {
+                if (!src) return Promise.resolve();
+                // format before saving
+                //src = pxtc.format(src, 0).formatted;
 
-            if (!src) return Promise.resolve();
-            // format before saving
-            //src = pxtc.format(src, 0).formatted;
-
-            let mainPkg = pkg.mainEditorPkg();
-            let tsName = this.editorFile.getVirtualFileName();
-            Util.assert(tsName != this.editorFile.name);
-            return mainPkg.setContentAsync(tsName, src).then(() => {
-                if (open) {
-                    let f = mainPkg.files[tsName];
-                    this.setFile(f);
-                }
+                let mainPkg = pkg.mainEditorPkg();
+                let tsName = this.editorFile.getVirtualFileName();
+                Util.assert(tsName != this.editorFile.name);
+                return mainPkg.setContentAsync(tsName, src).then(() => {
+                    if (open) {
+                        let f = mainPkg.files[tsName];
+                        this.setFile(f);
+                    }
+                });
             });
         });
