+++ webapp/src/tdlegacy.ts
@@ -1,14 +1,13 @@
 import * as core from "./core";
-import * as workeriface from "./workeriface"
 import * as pkg from "./package";
 
 import Cloud = pxt.Cloud;
 import U = pxt.Util;
 
-let iface: workeriface.Iface
+let iface: pxt.worker.Iface
 
 export function td2tsAsync(td: string) {
-    if (!iface) iface = workeriface.makeWebWorker(pxt.webConfig.tdworkerjs)
+    if (!iface) iface = pxt.worker.makeWebWorker(pxt.webConfig.tdworkerjs)
 
     return pkg.mainPkg.getCompileOptionsAsync()
         .then(opts => {
