+++ worker/worker.ts
@@ -1,5 +1,6 @@
-/// <reference path="../../built/pxtlib.d.ts"/>
-/// <reference path="../../built/pxttypescript.d.ts"/>
+/// <reference path="../built/pxtlib.d.ts"/>
+/// <reference path="../built/pxttypescript.d.ts"/>
+
 importScripts(
     "/blb/typescript.js",
     "/blb/fuse.min.js",
